backend/sistema_nomina/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
et_xmlfile==2.0.0
num2words==0.5.14
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
